refactor #18 : model에 따른 type 분리
