+++ src/models/common/index.ts
@@ -1,0 +1,1 @@
+export type NonEmptyArray<T> = readonly [T, ...T[]];
